Refactor
Move functions around to clean up `app.py`
