--- tempCodeRunnerFile.py
@@ -1,2 +1,0 @@
-
-    advanced_white_games = get_advanced_white_games(chess_games)
